PSD-1039: Move elasticsearch-related things on investigation to a separate module
